requirements.txt
selenium==4.16.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
pillow==10.1.0
pytesseract==0.3.10
openai>=2.0.0
python-dateutil==2.8.2
piexif==1.1.3
